Web Application - sources/app/requirements.txt
flask
flask[async]
python-dotenv
gunicorn
pydantic
pydantic_settings
apscheduler
